backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pymssql>=2.3.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
pydantic[email]>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
cryptography>=42.0.0
reportlab>=4.2.0
pillow>=11.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
